userLogin, showUser ejs complete
