+++ TODO.md
@@ -1,0 +1,4 @@
+Reviews - Need to add a reviews controller, model, and view seperate from user
+    Need to add to user after these are created
+
+Authentication- need to add authentication to login
